Return the commission as a whole number
This also addresses some feedback from @pepopowitz around the signal and
noise in the query spec. Rather than querying for things that don't
matter, I removed the outer stuff about the submission. Then I used an
object matcher to assert about all the things I was querying for on the
offer.

And this was really valuable because it revealed a bug! I was passing a
float into an integer and getting wacky results in Volt. I now see what
I really wanted to do was request the whole number and then work with it
like that. So yeah, bravo for great PR feedback @pepopowitz!!! :tada: